Adding work from last week and a half.
Worked on setting up a test folder for the RH data, converting RH data to an xts object, computing daily returns for the respective stocks, and finding the difference in RH participation. Still not sure about the difference in RH participation. The time stamps on the RH data were hourly so I wasn't sure how to compute the daily difference.